Renamed the base class appropriately so that we can create a baser class without menu options
This class will be meant for utilities that do not have the actionBar. Because of the lack of it, there isn't really a need to initialize an options menu or even attempt to as it will bork some layouts

Therefore we will remove it appropirately

With this we should be just about done in implementing this utility. So
this commit will officially close #281