Fixed bug where semesters in semester list is sorted one way, but not actually
Basically the bug is that when you click on the semester that has been sorted a single way, you may get a different semester when you enter the module view. This is because only the semester is sorted but the keys are unsorted still. As a result there was a discrepency. This commit resolves that discrepency and restores the intended operation