[Cloud Functions] Added GPA Ordering and async setRecord function
We are going to handle ordering based on 2 new values (not implemented as of now in the Android app) where it will take require a start and end time for the institution to order. If it is not found it will be generated for the server side
In the app side the same default values (currentTimeMillis for start and -1 for end) will be found too. This was added on the server to prevent old data crashing

Finally the order will be based off either the start or end timestamp depending on if there is a presence of the end timestamp.

Also we updated the setRecord function to make use of async/await (Node 8) too and cleaned up some logs