CUTILAND-368 Fixed a bug where successful authentication is treated as cancelled
This is apparently caused by the DayNight app handling restarting the activity when changing themes due to it invoking a configuration change (and hence activity restart) when the theme changes. Due to this it actually invoked the authentication activity twice (1 cancelled, and 1 waiting user input). When the user authenticate and causes a return, the authentication activity actually returns the cancelled one first, causing the calling activity to exit due to failure in authentication

This CR fixes it by making the theme change check appear BEFORE the activity is created so that double configuration changes does not occur

Fixes CUTILAND-368

Signed-off-by: itachi1706 <kennethsohyq@gmail.com>