CUTILAND-389 Defer screen lock non deprecated implementation to Android 10 targetSDK
We also renamed the prompt a little to make it better (like wth is Sign In hahahah)

It does not reliably work in API 28, clicking Use PIN or Use Pattern will result in authentication just cancelling the authentication completely with no results being returned with the secondary auth, making the app stuck at the authentication screen. Reports from reddit, stackoverflow and the Android Bug Tracker (See a/145231213 and a/142740104) shows that it is apparently an issue with Android 10 devices. So for now we will also retain and return to this when the bug has been reliabily fixed. Hence we forked the issue out to a seperate issue to be handled when we need to

Resolves CUTILAND-389 (for now)