CUTILAND-408 Implemented Camera Viewer utility
This is previously tested and implemented based off of https://gitlab.com/itachi1706/android-camera2-sample

Note: This utility requires devices to be API 21 and above as it makes heavy use of the Android Camera2 API, which is only available from Android Lollipop devices onwards. However as this is part of the sprint with the menu revamp that will bring the feature, there are no checks to ensure that users are on a supported Android version for this utility. Note that if you attempt to launch the utility on a pre-Lollipop device, the application WILL crash

Resolves CUTILAND-408